Create new function for uniq date, change derictory for disactive user, create route singup
